move VcrActor to own file, start actor with supervision, add some spec helpers
