+++ spec/vcr/library_hooks/webmock_spec.rb
@@ -38,7 +38,7 @@ describe "WebMock hook", :with_monkey_patches => :webmock do
       expect(webmock_request.instance_variables.map(&:to_sym)).not_to include(:@__typed_vcr_request)
     end
 
-    context "when there'ss a bug and the request does not have the @__typed_vcr_request in the after_request callbacks" do
+    context "when there's a bug and the request does not have the @__typed_vcr_request in the after_request callbacks" do
       let(:warner) { VCR::LibraryHooks::WebMock }
       before { allow(warner).to receive(:warn) }
